UpdateResponsable

+++ src/views/responsables/index.php
@@ -11,9 +11,11 @@
       <div class="card">
         <div class="card-body">
           <div class="card-title text-center fs-3 text-success fw-bold">Responsable</div>
-          <a href="index.php?controller=responsable&action=add" class="btn btn-dark mb-3">
-            Nouveau responsable
-          </a>
+          <?php if ($_SESSION['user']['role'] === 'ADMIN'): ?>
+            <a href="index.php?controller=responsable&action=add" class="btn btn-dark mb-3">
+              Nouveau responsable
+            </a>
+          <?php endif; ?>
 
           <table class="table table-bordered border border-1 rounded-3">
             <tr>
@@ -24,19 +26,21 @@
             </tr>
 
             <?php foreach ($responsables as $r): ?>
-            <tr>
-              <td class="fs-5 "><?= $r['nomresponsable'] ?> <?= $r['prenomresponsable'] ?></td>
-              <td class="fs-5 "><?= $r['mailresponsable'] ?></td>
-              <td class="table-active fw-bold"><?= $r['role'] ?></td>
-              <td class="d-flex gap-5">
-                <a href="index.php?controller=responsable&action=edit&id=<?= $r['idresponsable'] ?>"
-                  class="btn btn-warning btn-sm">✏️</a>
+              <tr>
+                <td class="fs-5 "><?= $r['nomresponsable'] ?>   <?= $r['prenomresponsable'] ?></td>
+                <td class="fs-5 "><?= $r['mailresponsable'] ?></td>
+                <td class="table-active fw-bold"><?= $r['role'] ?></td>
+                <?php if ($_SESSION['user']['role'] === 'ADMIN'): ?>
+                  <td class="d-flex gap-5">
+                    <a href="index.php?controller=responsable&action=edit&id=<?= $r['idresponsable'] ?>"
+                      class="btn btn-warning btn-sm">✏️</a>
 
-                <a onclick="return confirm('Supprimer ?')"
-                  href="index.php?controller=responsable&action=delete&id=<?= $r['idresponsable'] ?>"
-                  class="btn btn-danger btn-sm">🗑️</a>
-              </td>
-            </tr>
+                    <a onclick="return confirm('Supprimer ?')"
+                      href="index.php?controller=responsable&action=delete&id=<?= $r['idresponsable'] ?>"
+                      class="btn btn-danger btn-sm">🗑️</a>
+                  </td>
+                <?php endif; ?>
+              </tr>
             <?php endforeach; ?>
           </table>
         </div>
